Refactoring the BaseEndOfTurnAnalyzer to include most of the logic
